requirements.txt
PyQt6>=6.5.0
Pillow>=10.0.0
numpy>=1.24.0
reportlab>=4.0.0
scipy>=1.11.0